# More optimizations for AES GCM and CCM