The PR implements GHASH with small table, in case `support_pclmuldq` is not possible. There is a slowdown on the pclmuldq benchmark, but it comes from increasing AESGCM context size and `memory` scrubbing 1 byte at a time instead of using machine words. This can be fixed with a `memory` PR.

CCM is also improved. It was especially slow because not calling the AESNI implementation on the authentication part.

cryptonite-0.26 without support_pclmuldq:
```
AE/ChaChaPoly1305                        mean 7.644 μs  ( +- 188.5 ns  )
AE/AES-GCM                               mean 14.47 μs  ( +- 78.51 ns  )
AE/AES-CCM                               mean 21.60 μs  ( +- 109.4 ns  )
```

cryptonite-0.26 with support_pclmuldq:
```
AE/ChaChaPoly1305                        mean 7.620 μs  ( +- 145.4 ns  )
AE/AES-GCM                               mean 3.018 μs  ( +- 45.63 ns  )
AE/AES-CCM                               mean 21.62 μs  ( +- 89.81 ns  )
```

PR without support_pclmuldq:
```
AE/ChaChaPoly1305                        mean 7.782 μs  ( +- 174.3 ns  )
AE/AES-GCM                               mean 7.689 μs  ( +- 136.2 ns  )
AE/AES-CCM                               mean 3.983 μs  ( +- 41.11 ns  )
```

PR with support_pclmuldq:
```
AE/ChaChaPoly1305                        mean 7.649 μs  ( +- 150.8 ns  )
AE/AES-GCM                               mean 3.467 μs  ( +- 139.2 ns  )
AE/AES-CCM                               mean 3.968 μs  ( +- 61.69 ns  )
```